backend/regulatory_compliance/requirements.txt
django==5.1.1
djangorestframework==3.15.2
pymongo==4.8.0
psycopg2-binary==2.9.9
PyPDF2==3.0.1
google-generativeai==0.8.3